pyproject.toml
[tool.poetry]
name = "pismo-tech-case"
version = "0.1.0"
description = ""
authors = [
    "LeeoSilva <leeosilva0909@gmail.com>"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.16.1"
sqlalchemy = "^2.0.41"
fastapi = {extras = ["all"], version = "^0.115.12"}
pydantic = "^2.11.5"
sqlmodel = "^0.0.24"
gunicorn = "^23.0.0"
python-slugify = "^8.0.4"
pydantic-settings = "^2.9.1"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-cov = "^6.1.1"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
